app/AlamLab/requirements.txt
pyserial
pyside6